requirements.txt
fastapi~=0.92.0
uvicorn==0.20.0
pymongo~=4.3.3
pydantic~=1.10.5
motor~=3.1.1
pre-commit==2.17.0
flake8==4.0.1
flake8-isort==4.1.1
mypy==0.931
cryptography==38.0.4
python-jose==3.3.0
passlib==1.7.4
